+++ libarchive/archive_write_add_filter_compress.c
@@ -397,11 +397,11 @@ archive_compressor_compress_write(struct archive_write_filter *f,
 		state->checkpoint = state->in_count + CHECK_GAP;
 
 		if (state->in_count <= 0x007fffff)
-			ratio = state->in_count * 256 / state->out_count;
-		else if ((ratio = state->out_count / 256) == 0)
+			ratio = (int)(state->in_count * 256 / state->out_count);
+		else if ((ratio = (int)(state->out_count / 256)) == 0)
 			ratio = 0x7fffffff;
 		else
-			ratio = state->in_count / ratio;
+			ratio = (int)(state->in_count / ratio);
 
 		if (ratio > state->compress_ratio)
 			state->compress_ratio = ratio;
